+++ src/compas/datastructures/__init__.py
@@ -157,6 +157,8 @@ from .assembly.exceptions import AssemblyError, FeatureError
 from .assembly.assembly import Assembly
 from .assembly.part import Feature, GeometricFeature, ParametricFeature, Part
 
+from .tree.tree import Tree, TreeNode
+
 BaseNetwork = Network
 BaseMesh = Mesh
 BaseVolMesh = VolMesh
@@ -274,6 +276,9 @@ __all__ = [
     "Feature",
     "GeometricFeature",
     "ParametricFeature",
+    # Trees
+    "Tree",
+    "TreeNode",
 ]
 
 if not compas.IPY: