+++ src/compas/datastructures/tree/tree.py
@@ -3,30 +3,53 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import annotations
 
-from compas.data import Data
 from compas.datastructures import Datastructure
 
 
-class TreeNode(Data):
-    JSONSCHEMA = {}
+class TreeNode(object):
+    """A node of a tree data structure.
+
+    Parameters
+    ----------
+    name: str, optional
+        The name of the tree ndoe.
+    attributes: dict[str, Any], optional
+        User-defined attributes of the datastructure.
+
+    Attributes
+    ----------
+    name: str
+        The name of the datastructure.
+    attributes : dict[str, Any]
+        User-defined attributes of the datastructure.
+    parent: :class:`~compas.datastructures.TreeNode`
+        The parent node of the tree node.
+    children: set[:class:`~compas.datastructures.TreeNode`]
+        The children of the tree node.
+    tree: :class:`~compas.datastructures.Tree`
+        The tree the node belongs to.
+    is_root: bool
+        True if the node is the root node of the tree.
+    is_leaf: bool
+        True if the node is a leaf node of the tree.
+    is_branch: bool
+        True if the node is a branch node of the tree.
+    acestors: generator[:class:`~compas.datastructures.TreeNode`]
+        A generator of the acestors of the tree node.
+    descendants: generator[:class:`~compas.datastructures.TreeNode`]
+        A generator of the descendants of the tree node.
+
+    """
 
     def __init__(self, name: str = None, attributes: dict = None):
-        super(TreeNode, self).__init__(name=name)
+        self.name = name
         self.attributes = attributes or {}
         self._parent = None
         self._children = set()
+        self._tree = None
 
     def __repr__(self) -> str:
-        return "TreeNode({})".format(self.name)
-
-    @property
-    def data(self):
-        return {
-            "name": self.name,
-            "attributes": self.attributes,
-            "parent": str(self.parent.guid) if self.parent else None,
-            "children": [str(child.guid) for child in self.children],
-        }
+        return "<TreeNode {}>".format(self.name)
 
     @property
     def is_root(self):
@@ -46,17 +69,29 @@ class TreeNode(Data):
 
     @property
     def children(self):
-        return list(self._children)
-
-    def add(self, node: TreeNode):
-        assert isinstance(node, TreeNode), "The node is not a TreeNode object."
-        self._children.add(node)
-        node._parent = self
+        return self._children
 
     @property
+    def tree(self):
+        return self._tree
+
+    def add(self, node: TreeNode):
+        """Add a child node to this node."""
+        if not isinstance(node, TreeNode):
+            raise TypeError("The node is not a TreeNode object.")
+        self._children.add(node)
+        node._parent = self
+        node._tree = self.tree
+        if self.tree:
+            self.tree.nodes.add(node)
+
     def remove(self, node: TreeNode):
+        """Remove a child node from this node."""
         self._children.remove(node)
         node._parent = None
+        node._tree = None
+        if self.tree:
+            self.tree.nodes.remove(node)
 
     @property
     def acestors(self):
@@ -73,38 +108,112 @@ class TreeNode(Data):
                 yield descendant
 
     def traverse(self):
+        """Traverse the tree from this node."""
         yield self
         for descendant in self.descendants:
             yield descendant
 
 
 class Tree(Datastructure):
+    """A tree data structure.
+
+    Parameters
+    ----------
+    name: str, optional
+        The name of the datastructure.
+    attributes: dict[str, Any], optional
+        User-defined attributes of the datastructure.
+
+    Attributes
+    ----------
+    name: str
+        The name of the datastructure.
+    attributes : dict[str, Any]
+        User-defined attributes of the datastructure.
+    root: :class:`~compas.datastructures.TreeNode`
+        The root node of the tree.
+    nodes: set[:class:`~compas.datastructures.TreeNode`]
+        The nodes of the tree.
+    leaves: generator[:class:`~compas.datastructures.TreeNode`]
+        A generator of the leaves of the tree.
+
+    Examples
+    --------
+    >>> from compas.datastructures import Tree, TreeNode
+    >>> tree = Tree()
+    >>> root = TreeNode('root')
+    >>> branch = TreeNode('branch')
+    >>> leaf1 = TreeNode('leaf1')
+    >>> leaf2 = TreeNode('leaf2')
+    >>> tree.add_root(root)
+    >>> root.add(branch)
+    >>> branch.add(leaf1)
+    >>> branch.add(leaf2)
+    >>> print(tree)
+    <Tree with 4 nodes, 1 branches, and 2 leaves>
+    >>> tree.print()
+    <TreeNode root>
+        <TreeNode branch>
+            <TreeNode leaf2>
+            <TreeNode leaf1>
+
+    """
+
     JSONSCHEMA = {}
 
-    def __init__(self):
+    def __init__(self, name: str = None, attributes: dict = None):
         super(Tree, self).__init__()
-        self.attributes = {}
+        self.name = name
+        self.attributes = attributes or {}
         self._root = None
+        self._nodes = set()
 
     @property
     def data(self):
+        def get_node_data(node):
+            return {
+                "name": node.name,
+                "attributes": node.attributes,
+                "children": [get_node_data(child) for child in node.children],
+            }
+
         return {
-            "root": str(self.root.guid) if self.root else None,
-            "nodes": [node.data for node in self.nodes],
+            "name": self.name,
+            "root": get_node_data(self.root),
+            "attributes": self.attributes,
         }
+
+    @data.setter
+    def data(self, data):
+        self.name = data["name"]
+        self.attributes = data["attributes"]
+
+        def node_from_data(data):
+            node = TreeNode(data["name"], data["attributes"])
+            for child in data["children"]:
+                node.add(node_from_data(child))
+            return node
+
+        self.add_root(node_from_data(data["root"]))
 
     @property
     def root(self):
         return self._root
 
     def add_root(self, node: TreeNode):
-        assert isinstance(node, TreeNode), "The node is not a TreeNode object."
+        """Add a root node to the tree."""
+        if not isinstance(node, TreeNode):
+            raise TypeError("The node is not a TreeNode object.")
         if not node.is_root:
             raise ValueError("The node is already part of another tree.")
         self._root = node
+        node._tree = self
+        self._nodes.add(node)
 
     def add(self, node: TreeNode, parent: TreeNode):
-        assert isinstance(node, TreeNode), "The node is not a TreeNode object."
+        """Add a node to the tree."""
+        if not isinstance(node, TreeNode):
+            raise TypeError("The node is not a TreeNode object.")
         if self.root is None:
             raise ValueError("The tree has no root node, use add_root() first.")
         else:
@@ -112,11 +221,18 @@ class Tree(Datastructure):
 
     @property
     def nodes(self):
-        return list(self.root.traverse())
+        return self._nodes
+
+    def remove_root(self):
+        """Remove the root node from the tree."""
+        self._root._tree = None
+        self.nodes.remove(self._root)
+        self._root = None
 
     def remove(self, node: TreeNode):
-        if node.is_root:
-            self._root = None
+        """Remove a node from the tree."""
+        if node == self.root:
+            self.remove_root()
         else:
             node.parent.remove(node)
 
@@ -126,40 +242,18 @@ class Tree(Datastructure):
             if node.is_leaf:
                 yield node
 
-    def summary(self):
+    def __repr__(self):
         nodes = len(self.nodes)
         branches = sum(1 for node in self.nodes if node.is_branch)
         leaves = sum(1 for node in self.nodes if node.is_leaf)
-        print("Tree with {} nodes, {} branches, and {} leaves".format(nodes, branches, leaves))
+        return "<Tree with {} nodes, {} branches, and {} leaves>".format(nodes, branches, leaves)
 
     def print(self):
+        """Print the spatial hierarchy of the tree."""
+
         def _print(node, depth=0):
             print("  " * depth + str(node))
             for child in node.children:
                 _print(child, depth + 1)
 
         _print(self.root)
-
-
-if __name__ == "__main__":
-    from compas.data import json_dumps
-
-    R = TreeNode("R")
-    B = TreeNode("B")
-    L = TreeNode("L")
-    L2 = TreeNode("L2")
-
-    T = Tree()
-    T.add_root(R)
-
-    # T.add(B, R)
-    # T.add(L, B)
-    # T.add(L2, B)
-
-    R.add(B)
-    B.add(L)
-    B.add(L2)
-
-    # T.print()
-
-    print(json_dumps(T, pretty=True))